make github token generation a util method
